requirements.txt
asgiref==3.5.2
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-restql==0.15.2
djangorestframework==3.13.1
drf-yasg==1.21.0
factory-boy==3.2.1
Faker==13.15.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.3
pyparsing==3.0.9
pyPEG2==2.15.2
python-dateutil==2.8.1
pytz==2022.1
requests==2.24.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.25.11
whitenoise==5.2.0
